cortex-agent-demo/requirements.txt
snowflake-connector-python>=3.6.0
requests>=2.31.0
